refactor: account registration service
