+++ src/main/java/fast/bank/api/domain/account/service/validation/registration/ExistingAndActiveUser.java
@@ -1,7 +1,6 @@
-package fast.bank.api.domain.account.service.registration.validation;
+package fast.bank.api.domain.account.service.validation.registration;
 
 import fast.bank.api.domain.account.dto.AccountRegistrationData;
-import fast.bank.api.domain.account.service.registration.validation.AccountRegistrationValidators;
 import fast.bank.api.domain.user.repository.UserRepository;
 import fast.bank.api.infra.exception.validation.ValidException;
 import org.springframework.beans.factory.annotation.Autowired;
